[core] New local variable concept: project type
* New file: core-local-variables.el
* New interactive function in Helm layer: spacemacs/helm-add-project-type

This new file expands the hack local variables capabilities of
Spacemacs with a new concept: Project Types

Each layer can add a project type with the function
`spacemacs/add-available-project-type`

With the function `` a user can add a type to the current project.

When a file is opened all the functions bound to the corresponding
project type using the function `spacemacs/add-local-var-hook` will
be executed (they are essentially setup functions to enable minor
modes conditionally).

The user can easily add a type for the current project with `SPC p s`
to set a type. It effectively adds the selected type to the `.dir-locals.el`
file at the root of the project.